// src/modules/sections/enums/SectionEnum.ts
export enum SectionTypeEnum {
  AUTHENTIC_AGENT = 'AUTHENTIC_AGENT', // Đại lý chính hãng
  TOP_AGENT = 'TOP_AGENT', // Đại lý nổi bật
  HOT_DEAL = 'HOT_DEAL', // Hot deal
  TOP_PRODUCT = 'TOP_PRODUCT', // Sản phẩm nổi bật
  BEST_SELLER_PRODUCT = 'BEST_SELLER_PRODUCT', // sản phẩm bán chạy
  MOST_VIEWED_PRODUCT = 'MOST_VIEWED_PRODUCT', // sản phảm tìm kiếm nhiều nhất
  MOST_VIEWED_SERVICE = 'MOST_VIEWED_SERVICE', // dịch vụ tìm kiếm nhiều nhất
  INSURANCE_PRODUCT = 'INSURANCE_PRODUCT', // bảo hiểm
  SPARE_PARTS_PRODUCT = 'SPARE_PARTS_PRODUCT', // phụ tùng phụ kiện
  TIRE_WHEEL_PRODUCT = 'TIRE_WHEEL_PRODUCT', // mâm lốp
  BATTERY_PRODUCT = 'BATTERY_PRODUCT', // ắc quy
  PROMOTION = 'PROMOTION', // khuyến mãi
  TOP_POST = 'TOP_POST', // bài viết nổi bật
}